refactor(protobuf): use message schema from definitions index
