+++ packages/connect-explorer/src/data/methods/cardano/getNativeScriptHash.ts
@@ -1,4 +1,4 @@
-import { CardanoNativeScriptType } from '@trezor/protobuf/src/messages-schema';
+import { CardanoNativeScriptType } from '@trezor/protobuf/src/definitions';
 
 import { cardanoDerivationType, cardanoNativeScriptHashDisplayFormat } from './common';
