+++ packages/connect-explorer/src/pages/details/protobuf.mdx
@@ -1,4 +1,4 @@
-import { MessageType } from '@trezor/protobuf/src/messages-schema';
+import { MessageType } from '@trezor/protobuf/src/definitions';
 
 import { ParamsTable } from '../../components/ParamsTable';
