+++ packages/connect-explorer/src/pages/methods/bitcoin/signMessage.mdx
@@ -46,7 +46,7 @@ TrezorConnect.signMessage({
 
 ### Result
 
-[MessageSignature type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/messages.ts)
+[MessageSignature type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/definitions/messages-bitcoin.ts)
 
 ```javascript
 {
