+++ packages/connect-explorer/src/pages/methods/device/getFirmwareHash.mdx
@@ -50,7 +50,7 @@ TrezorConnect.getFirmwareHash({
 
 ### Result
 
-[FirmwareHash type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/messages.ts)
+[FirmwareHash type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/definitions/messages-management.ts)
 
 ```javascript
 {
