+++ packages/connect-explorer/src/pages/methods/ethereum/ethereumSignMessage.mdx
@@ -49,7 +49,7 @@ TrezorConnect.ethereumSignMessage({
 
 ### Result
 
-[MessageSignature type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/messages.ts)
+[MessageSignature type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/definitions/messages-bitcoin.ts)
 
 ```javascript
 {
