+++ packages/connect-explorer/src/pages/methods/ethereum/ethereumSignTypedData.mdx
@@ -114,7 +114,7 @@ TrezorConnect.ethereumSignTypedData({
 
 ### Result
 
-[EthereumMessageSignature type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/messages.ts)
+[EthereumMessageSignature type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/definitions/messages-ethereum.ts)
 
 ```javascript
 {
