+++ packages/connect-explorer/src/pages/methods/tezos/tezosSignTransaction.mdx
@@ -233,7 +233,7 @@ TrezorConnect.tezosSignTransaction({
 
 ### Result
 
-[TezosSignedTx type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/messages.ts)
+[TezosSignedTx type](https://github.com/trezor/trezor-suite/blob/develop/packages/protobuf/src/definitions/messages-tezos.ts)
 
 ```javascript
 {
