+++ packages/connect/src/device/__tests__/checkFirmwareHash.test.ts
@@ -1,6 +1,6 @@
 import type { ConnectSettings, DeviceUniquePath } from '@trezor/connect-common';
 import { DeviceModelInternal } from '@trezor/device-utils';
-import type { FirmwareHash } from '@trezor/protobuf/src/messages';
+import type { FirmwareHash } from '@trezor/protobuf/src/definitions';
 import type { Descriptor } from '@trezor/transport';
 import { Log } from '@trezor/utils';
