+++ packages/connect/src/exports.ts
@@ -1,7 +1,7 @@
 export * from '@trezor/connect-common';
 
 // note: only @trezor/connect re-exports protobuf runtime
-export { Messages as PROTO } from '@trezor/protobuf';
+export { MessagesSchema as PROTO } from '@trezor/protobuf';
 // note: backwards compatibility with suite. I don't think we need to export this publicaly
 export { TRANSPORT } from '@trezor/transport';
